Integration pixellation and trochoid shaders
Trochoid is a WIP. Need to do some magic to make it properly shear. In
short: it's currently implemented as a standalone Mesh Renderer object
which I place on my avatar's neck bone. Since its object origin is not
at the hip bone like everything else, it behaves weirdly when shearing.

Solution is to implement it as a regular skinned mesh renderer. Requires
some careful analysis to get right.