Decals can now use a secondary UV channel
This lets you efficiently create a shit ton of tattoos using a single
texture. It also lets you place them over seamlines. Just create a
secondary UV map in blender and export as FBX. The shader will pick up
this secondary channel and use it, if instructed to do so in the decals
UI.