Add small programmable text display gimmick
Use Third_Party/gen_atlas to create a 512x512 texture and put it in the
texture slot (the resolution is hardcoded to 512 in shader). Use sliders to
control the grid size. If desired, use global offset to make it possible
to linearly animate between e.g. 0-9 and have it display as ASCII
'0'-'9'.